backend/requirements.txt
fastapi
uvicorn
pydantic
requests
google-api-python-client
youtube-transcript-api
google-generativeai
python-dotenv
authlib
itsdangerous
python-jose[cryptography]
httpx